refactor: 회원가입 페이지 반환을 위한 doGet 메서드 구현
