+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -2,6 +2,8 @@ package com.techcourse.controller;
 
 import com.techcourse.model.User;
 import com.techcourse.service.AuthService;
+import com.techcourse.util.ResourceLoader;
+import java.io.IOException;
 import java.util.UUID;
 import org.apache.catalina.Session;
 import org.apache.catalina.SessionManager;
@@ -12,6 +14,12 @@ import org.apache.coyote.http11.response.HttpResponse;
 public class RegisterController extends AbstractController {
 
     private final AuthService authService = new AuthService();
+
+    @Override
+    protected HttpResponse doGet(HttpRequest request) throws IOException {
+        String path = request.getPath();
+        return HttpResponse.ok(ResourceLoader.readWithFallback(path), ResourceLoader.getMimeType(path));
+    }
 
     @Override
     protected HttpResponse doPost(HttpRequest request) {
